Crate alloy_rpc_types

alloy-rpc-types

Meta-crate for all Ethereum JSON-RPC types.

Re-exports

pub use alloy_serde as serde_helpers;

pub use alloy_rpc_types_admin as admin;

admin

pub use alloy_rpc_types_anvil as anvil;

anvil

pub use alloy_rpc_types_beacon as beacon;

beacon

pub use alloy_rpc_types_engine as engine;

engine

pub use alloy_rpc_types_eth as eth;

eth

pub use alloy_rpc_types_trace as trace;

trace

pub use alloy_rpc_types_txpool as txpool;

txpool

Modules

erroreth

Commonly used errors for the eth_ namespace.

pubsubeth

Ethereum types for pub-sub

stateeth

bindings for state overrides in eth_call

transactioneth

RPC types for transactions

Structs

AccessListeth

AccessList as defined in EIP-2930

AccessListItemeth

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListWithGasUsedeth

Access list with gas used appended.

AccountInfoeth

Account information.

AnyReceiptEnvelopeeth

Receipt envelope, as defined in EIP-2718.

BlobTransactionSidecareth

This represents a set of blobs, and its corresponding commitments and proofs.

Blocketh

Block representation

BlockNumHasheth

Block number and hash.

BlockOverrideseth

BlockOverrides is a set of header fields to override.

BlockTransactionHasheseth

An iterator over the transaction hashes of a block.

BlockTransactionHashesMuteth

An Iterator over the transaction hashes of a block.

BloomFiltereth

Helper type to represent a bloom filter used for matching logs.

Bundleeth

Bundle of transactions

ChainStatuseth

Chain status.

EIP1186AccountProofResponseeth

Response for EIP-1186 account proof eth_getProof

EIP1186StorageProofeth

Data structure with proof for one single storage-entry

EthCallResponseeth

CallResponse for eth_callMany

ExtAccountInfoeth

Extended account information (used by parity_allAccountInfo).

FeeHistoryeth

Response type for eth_feeHistory

Filtereth

Filter for logs.

FilterSeteth

FilterSet is a set of values that will be used to filter logs

FilteredParamseth

Support for matching Filters

Headereth

Block header representation.

Indexeth

A hex encoded or decimal index that’s intended to be used as a rust index, hence it’s deserialized into a usize.

Logeth

Ethereum Log emitted by a transaction

OptimismTransactionReceiptFieldseth

Additional fields for Optimism transaction receipts

Parityeth

Type that represents the signature parity byte, meant for use in RPC.

PeerEthProtocolInfoeth

Peer Ethereum protocol information

PeerInfoeth

Peer connection information

PeerNetworkInfoeth

Peer network information

PeerProtocolsInfoeth

Peer protocols information

Peerseth

Peers info

PipProtocolInfoeth

Peer PIP protocol information

RawLogeth

Ethereum Log

Receipteth

Receipt containing result of transaction execution.

ReceiptWithBloometh

Receipt with calculated bloom filter.

RecoveredAccounteth

account derived from a signature as well as information that tells if it is valid for the current chain

Richeth

Value representation with additional info

RpcBlockHasheth

A block hash which may have a boolean requireCanonical field. If false, an RPC call should raise if a block matching the hash is not found. If true, an RPC call should additionally raise if the block is not in the canonical chain. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md#specification

Signatureeth

Container type for all signature fields in RPC

StateContexteth

State context for callMany

SyncInfoeth

Syncing info

Transactioneth

Transaction object used in RPC

TransactionInfoeth

Additional fields in the context of a block that contains this transaction.

TransactionInputeth

Helper type that supports both data and input fields that map to transaction input data.

TransactionReceipteth

Transaction receipt

TransactionRequesteth

Represents all transaction requests to/from RPC.

TransactionStatseth

Propagation statistics for pending transaction.

TxGasAndRewardeth

Internal struct to calculate reward percentiles

Withdrawaleth

Withdrawal represents a validator withdrawal from the consensus layer.

Worketh

The result of an eth_getWork request

Enums

BlockErroreth

Error that can occur when converting other types to blocks

BlockHashOrNumbereth

Either a block hash or a block number

BlockIdeth

A Block Identifier. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTageth

A block Number (or tag - “latest”, “earliest”, “pending”)

BlockTransactionseth

Block Transactions depending on the boolean attribute of eth_getBlockBy*, or if used by eth_getUncle*

BlockTransactionsKindeth

Determines how the transactions field of Block should be filled.

ConversionErroreth

Error variants when converting from crate::Transaction to alloy_consensus::Signed transaction.

FilterBlockOptioneth

Represents the target range of blocks for the filter

FilterChangeseth

Response of the eth_getFilterChanges RPC.

FilterIdeth

Owned equivalent of a SubscriptionId

PeerCounteth

Number of peers connected to.

PendingTransactionFilterKindeth

Specifies the kind of information you wish to receive from the eth_newPendingTransactionFilter RPC endpoint.

ReceiptEnvelopeeth

Receipt envelope, as defined in EIP-2718.

SyncStatuseth

Sync status

ValueOrArrayeth

Union type for representing a single value or a vector of values inside a filter

Functions

calc_blob_gaspriceeth

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gaseth

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.

logs_bloometh

Calculate receipt logs bloom.

Type Aliases

AnyTransactionReceipteth

Alias for a catch-all receipt type.

ForkBlocketh

Block number and hash of the forked block.

RichBlocketh

A Block representation that allows to include additional fields

RichHeadereth

Header representation with additional info.

Topiceth

A single topic