Struct alloy_rpc_types::state::EvmOverrides

pub struct EvmOverrides {
    pub state: Option<HashMap<Address, AccountOverride, FbBuildHasher<20>>>,
    pub block: Option<Box<BlockOverrides>>,
}

Helper type that bundles various overrides for EVM Execution.

By Default, no overrides are included.

Fields
state: Option<HashMap<Address, AccountOverride, FbBuildHasher<20>>>

Applies overrides to the state before execution.

block: Option<Box<BlockOverrides>>

Applies overrides to the block before execution.

This is a Box because less common and only available in debug trace endpoints.

Implementations

impl EvmOverrides

pub const fn new( state: Option<HashMap<Address, AccountOverride, FbBuildHasher<20>>>, block: Option<Box<BlockOverrides>>, ) -> EvmOverrides

Creates a new instance with the given overrides

pub const fn state( state: Option<HashMap<Address, AccountOverride, FbBuildHasher<20>>>, ) -> EvmOverrides

Creates a new instance with the given state overrides.

pub const fn block(block: Option<Box<BlockOverrides>>) -> EvmOverrides

Creates a new instance with the given block overrides.

pub const fn has_state(&self) -> bool

Returns true if the overrides contain state overrides.

pub const fn has_block(&self) -> bool

Returns true if the overrides contain block overrides.

pub fn with_state( self, state: HashMap<Address, AccountOverride, FbBuildHasher<20>>, ) -> EvmOverrides

Adds state overrides to an existing instance.

pub fn with_block(self, block: Box<BlockOverrides>) -> EvmOverrides

Adds block overrides to an existing instance.