Struct alloy_rpc_types::Block

pub struct Block<T = Transaction, H = Header> {
    pub header: H,
    pub uncles: Vec<FixedBytes<32>>,
    pub transactions: BlockTransactions<T>,
    pub withdrawals: Option<Withdrawals>,
}

Block representation for RPC.

Fields
header: H

Header of the block.

uncles: Vec<FixedBytes<32>>

Uncles’ hashes.

transactions: BlockTransactions<T>

Block Transactions. In the case of an uncle block, this field is not included in RPC responses, and when deserialized, it will be set to BlockTransactions::Uncle.

withdrawals: Option<Withdrawals>

Withdrawals in the block.

Implementations

impl<T, H> Block<T, H>

pub const fn empty(header: H) -> Block<T, H>

Creates a new empty block (without transactions).

pub const fn new(header: H, transactions: BlockTransactions<T>) -> Block<T, H>

Creates a new Block with the given header and transactions.

Note: This does not set the withdrawals for the block.

use alloy_eips::eip4895::Withdrawals;
use alloy_network_primitives::BlockTransactions;
use alloy_rpc_types_eth::{Block, Header, Transaction};
let block = Block::new(
    Header::new(alloy_consensus::Header::default()),
    BlockTransactions::<Transaction>::Full(vec![]),
)
.with_withdrawals(Some(Withdrawals::default()));

pub fn with_transactions( self, transactions: BlockTransactions<T>, ) -> Block<T, H>

Sets the transactions for the block.

pub fn with_withdrawals(self, withdrawals: Option<Withdrawals>) -> Block<T, H>

Sets the withdrawals for the block.

pub fn with_uncles(self, uncles: Vec<FixedBytes<32>>) -> Block<T, H>

Sets the uncles for the block.

pub fn map_header<U>(self, f: impl FnOnce(H) -> U) -> Block<T, U>

Converts the block’s header type by applying a function to it.

pub fn try_map_header<U, E>( self, f: impl FnOnce(H) -> Result<U, E>, ) -> Result<Block<T, U>, E>

Converts the block’s header type by applying a fallible function to it.

pub fn map_transactions<U>(self, f: impl FnMut(T) -> U) -> Block<U, H>

Converts the block’s transaction type by applying a function to each transaction.

Returns the block with the new transaction type.

pub fn try_map_transactions<U, E>( self, f: impl FnMut(T) -> Result<U, E>, ) -> Result<Block<U, H>, E>

Converts the block’s transaction type by applying a fallible function to each transaction.

Returns the block with the new transaction type if all transactions were successfully.

impl<T, H> Block<T, H>
where T: TransactionResponse,

pub fn into_full_block(self, txs: Vec<T>) -> Block<T, H>

Converts a block with Tx hashes into a full block.

impl<T, H> Block<T, Header<H>>
where H: Sealable + Encodable,

pub fn uncle_from_header(header: H) -> Block<T, Header<H>>

Constructs an “uncle block” from the provided header.

This function creates a new Block structure for uncle blocks (ommer blocks), using the provided alloy_consensus::Header.

impl<T> Block<T>

pub fn from_consensus( block: Block<T>, total_difficulty: Option<Uint<256, 4>>, ) -> Block<T>
where T: Encodable,

Constructs block from a consensus block and total_difficulty.

pub fn into_consensus(self) -> Block<T>

Consumes the block and returns the alloy_consensus::Block.

This has two caveats:

The returned block will always have empty uncles.

If the block’s transaction is not BlockTransactions::Full, the returned block will have an empty transaction vec.