Struct alloy_rpc_types::TransactionReceipt

pub struct TransactionReceipt<T = ReceiptEnvelope<Log>> {

    pub inner: T,
    pub transaction_hash: FixedBytes<32>,
    pub transaction_index: Option<u64>,
    pub block_hash: Option<FixedBytes<32>>,
    pub block_number: Option<u64>,
    pub gas_used: u128,
    pub effective_gas_price: u128,
    pub blob_gas_used: Option<u128>,
    pub blob_gas_price: Option<u128>,
    pub from: Address,
    pub to: Option<Address>,
    pub contract_address: Option<Address>,
    pub state_root: Option<FixedBytes<32>>,

}

Available on crate feature eth only.

Transaction receipt

This type is generic over an inner ReceiptEnvelope which contains consensus data and metadata.

Fields
inner: T

The receipt envelope, which contains the consensus receipt data..

transaction_hash: FixedBytes<32>

Transaction Hash.

transaction_index: Option<u64>

Index within the block.

block_hash: Option<FixedBytes<32>>

Hash of the block this transaction was included within.

block_number: Option<u64>

Number of the block this transaction was included within.

gas_used: u128

Gas used by this transaction alone.

effective_gas_price: u128

The price paid post-execution by the transaction (i.e. base fee + priority fee). Both fields in 1559-style transactions are maximums (max fee + max priority fee), the amount that’s actually paid by users can only be determined post-execution

blob_gas_used: Option<u128>

Blob gas used by the eip-4844 transaction

This is None for non eip-4844 transactions

blob_gas_price: Option<u128>

The price paid by the eip-4844 transaction per blob gas.

from: Address

Address of the sender

to: Option<Address>

Address of the receiver. None when its a contract creation transaction.

contract_address: Option<Address>

Contract address created, or None if not a deployment.

state_root: Option<FixedBytes<32>>

The post-transaction stateroot (pre Byzantium)

EIP98 makes this optional field, if it’s missing then skip serializing it

Implementations

impl TransactionReceipt

pub const fn status(&self) -> bool

Returns the status of the transaction.

pub const fn transaction_type(&self) -> TxType

Returns the transaction type.

pub fn calculate_create_address(&self, nonce: u64) -> Option<Address>

Calculates the address that will be created by the transaction, if any.

Returns None if the transaction is not a contract creation (the to field is set), or if the from field is not set.

impl<T> TransactionReceipt<T>

pub fn map_inner<U, F>(self, f: F) -> TransactionReceipt<U>
where F: FnOnce(T) -> U,

Maps the inner receipt value of this receipt.