Type Alias alloy_rpc_types::RichHeader

pub type RichHeader = Rich<Header>;

Available on crate feature eth only.

Header representation with additional info.

Aliased Type

struct RichHeader {
    pub inner: Header,
    pub extra_info: BTreeMap<String, Value>,
}

Fields
inner: Header

Standard value.

extra_info: BTreeMap<String, Value>

Additional fields that should be serialized into the Block object

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 912 bytes