Module alloy_rpc_types::eth::transaction

RPC types for transactions

Re-exports

pub use common::TransactionInfo;

pub use error::ConversionError;

pub use optimism::OptimismTransactionReceiptFields;

pub use receipt::AnyTransactionReceipt;

pub use receipt::TransactionReceipt;

pub use request::TransactionInput;

pub use request::TransactionRequest;

pub use signature::Parity;

pub use signature::Signature;

Modules

common 🔒

Commonly used additional types that are not part of the JSON RPC spec but are often required when working with RPC types, such as Transaction

error 🔒

optimism

Misc Optimism-specific types.

receipt 🔒

request

Alloy basic Transaction Request type.

signature 🔒

Signature related RPC values

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListWithGasUsed

Access list with gas used appended.

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Transaction

Transaction object used in RPC

Enums

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.