alloy_rpc_types

Module simulate

‘eth_simulateV1’ Request / Response types: https://github.com/ethereum/execution-apis/pull/484

Structs

SimBlock

Represents a batch of calls to be simulated sequentially within a block. This struct includes block and state overrides as well as the transaction requests to be executed.

SimCallResult

Captures the outcome of a transaction simulation. It includes the return value, logs produced, gas used, and the status of the transaction.

SimulateError

The error response returned by the eth_simulateV1 method.

SimulatePayload

Simulation options for executing multiple blocks and transactions.

SimulatedBlock

Represents the result of simulating a block.

Constants

MAX_SIMULATE_BLOCKS

The maximum number of blocks that can be simulated in a single request,