Struct alloy_rpc_types::TransactionRequest

pub struct TransactionRequest {

    pub from: Option<Address>,
    pub to: Option<TxKind>,
    pub gas_price: Option<u128>,
    pub max_fee_per_gas: Option<u128>,
    pub max_priority_fee_per_gas: Option<u128>,
    pub max_fee_per_blob_gas: Option<u128>,
    pub gas: Option<u64>,
    pub value: Option<Uint<256, 4>>,
    pub input: TransactionInput,
    pub nonce: Option<u64>,
    pub chain_id: Option<u64>,
    pub access_list: Option<AccessList>,
    pub transaction_type: Option<u8>,
    pub blob_versioned_hashes: Option<Vec<FixedBytes<32>>>,
    pub sidecar: Option<BlobTransactionSidecar>,
    pub authorization_list: Option<Vec<SignedAuthorization>>,

}

Represents all transaction requests to/from RPC.

Fields
from: Option<Address>

The address of the transaction author.

to: Option<TxKind>

The destination address of the transaction.

gas_price: Option<u128>

The legacy gas price.

max_fee_per_gas: Option<u128>

The max base fee per gas the sender is willing to pay.

max_priority_fee_per_gas: Option<u128>

The max priority fee per gas the sender is willing to pay, also called the miner tip.

max_fee_per_blob_gas: Option<u128>

The max fee per blob gas for EIP-4844 blob transactions.

gas: Option<u64>

The gas limit for the transaction.

value: Option<Uint<256, 4>>

The value transferred in the transaction, in wei.

input: TransactionInput

Transaction data.

nonce: Option<u64>

The nonce of the transaction.

chain_id: Option<u64>

The chain ID for the transaction.

access_list: Option<AccessList>

An EIP-2930 access list, which lowers cost for accessing accounts and storages in the list. See EIP-2930 for more information.

transaction_type: Option<u8>

The EIP-2718 transaction type. See EIP-2718 for more information.

blob_versioned_hashes: Option<Vec<FixedBytes<32>>>

Blob versioned hashes for EIP-4844 transactions.

sidecar: Option<BlobTransactionSidecar>

Blob sidecar for EIP-4844 transactions.

authorization_list: Option<Vec<SignedAuthorization>>

Authorization list for for EIP-7702 transactions.

Implementations

impl TransactionRequest

pub const fn from(self, from: Address) -> TransactionRequest

Sets the from field in the call to the provided address

pub fn from_transaction<T>(tx: T) -> TransactionRequest
where T: Transaction,

Initializes the TransactionRequest with the provided transaction.

Note: This leaves the from field empty.

pub fn from_transaction_with_sender<T>( tx: T, from: Address, ) -> TransactionRequest
where T: Transaction,

Initializes the TransactionRequest with the provided transaction and sender.

pub const fn transaction_type(self, transaction_type: u8) -> TransactionRequest

Sets the transactions type for the transactions.

pub const fn gas_limit(self, gas_limit: u64) -> TransactionRequest

Sets the gas limit for the transaction.

pub const fn nonce(self, nonce: u64) -> TransactionRequest

Sets the nonce for the transaction.

pub const fn max_fee_per_gas(self, max_fee_per_gas: u128) -> TransactionRequest

Sets the maximum fee per gas for the transaction.

pub const fn max_priority_fee_per_gas( self, max_priority_fee_per_gas: u128, ) -> TransactionRequest

Sets the maximum priority fee per gas for the transaction.

pub const fn to(self, to: Address) -> TransactionRequest

Sets the recipient address for the transaction.

pub const fn value(self, value: Uint<256, 4>) -> TransactionRequest

Sets the value (amount) for the transaction.

pub fn access_list(self, access_list: AccessList) -> TransactionRequest

Sets the access list for the transaction.

pub fn input(self, input: TransactionInput) -> TransactionRequest

Sets the input data for the transaction.

pub fn fee_cap(&self) -> Option<u128>

Returns the configured fee cap, if any.

The returns gas_price (legacy) if set or max_fee_per_gas (EIP1559)

pub fn populate_blob_hashes(&mut self)

Populate the blob_versioned_hashes key, if a sidecar exists. No effect otherwise.

pub fn get_invalid_common_fields(&self) -> Vec<&'static str>

Gets invalid fields for all transaction types

pub fn get_invalid_1559_fields(&self) -> Vec<&'static str>

Gets invalid fields for EIP-1559 transaction type

pub fn trim_conflicting_keys(&mut self)

Trim field conflicts, based on the preferred type

This is used to ensure that the request will not be rejected by the server due to conflicting keys, and should only be called before submission via rpc.

pub const fn preferred_type(&self) -> TxType

Check this builder’s preferred type, based on the fields that are set.

Types are preferred as follows:

EIP-7702 if authorization_list is set

EIP-4844 if sidecar or max_blob_fee_per_gas is set

EIP-2930 if access_list is set

Legacy if gas_price is set and access_list is unset

EIP-1559 in all other cases

pub fn missing_keys(&self) -> Result<TxType, (TxType, Vec<&'static str>)>

Check if all necessary keys are present to build a transaction.

Returns

Ok(type) if all necessary keys are present to build the preferred type.

Err((type, missing)) if some keys are missing to build the preferred type.

pub fn complete_4844(&self) -> Result<(), Vec<&'static str>>

Check if all necessary keys are present to build a 4844 transaction, returning a list of keys that are missing.

NOTE: sidecar must be present, even if blob_versioned_hashes is set.

pub fn complete_1559(&self) -> Result<(), Vec<&'static str>>

Check if all necessary keys are present to build a 1559 transaction, returning a list of keys that are missing.

pub fn complete_2930(&self) -> Result<(), Vec<&'static str>>

Check if all necessary keys are present to build a 2930 transaction, returning a list of keys that are missing.

pub fn complete_7702(&self) -> Result<(), Vec<&'static str>>

Check if all necessary keys are present to build a 7702 transaction, returning a list of keys that are missing.

pub fn complete_legacy(&self) -> Result<(), Vec<&'static str>>

Check if all necessary keys are present to build a legacy transaction, returning a list of keys that are missing.

pub fn buildable_type(&self) -> Option<TxType>

Return the tx type this request can be built as. Computed by checking the preferred type, and then checking for completeness.

pub fn build_typed_tx(self) -> Result<TypedTransaction, TransactionRequest>

Build a TypedTransaction

When Ok(...) is returned, the TypedTransaction is guaranteed to be complete. Which is to say, that it is signable, and the signed versino can be sent to the network.

pub fn build_consensus_tx(self) -> Result<TypedTransaction, BuildTransactionErr>

Build a TypedTransaction.

When Ok(...) is returned, the TypedTransaction is not guaranteed to be complete, only signable.

E.g. a particular case is when the transaction is of type Eip4844 and the sidecar is not set, in this case the transaction is not complete, i.e. it cannot be sent to the network once signed. However, it can still be used to calculate the signing hash, signature of the transaction, and transaction trie hash.

In case the requirement is to build a complete transaction, use build_typed_tx instead.